feat: adjust analysis-bucket handling across services

+++ packages/server-analysis-manager/src/components/core/commands/destroy/module.ts
@@ -25,14 +25,16 @@ export async function executeCoreDestroyCommand(
     for (let i = 0; i < buckets.length; i++) {
         const bucket = buckets[i];
 
-        try {
-            await storage.bucket.delete(bucket.external_id);
-        } catch (e) {
-            if (isClientErrorWithStatusCode(e, [404])) {
-                continue;
-            }
+        if(bucket.external_id) {
+            try {
+                await storage.bucket.delete(bucket.external_id);
+            } catch (e) {
+                if (isClientErrorWithStatusCode(e, [404])) {
+                    continue;
+                }
 
-            throw e;
+                throw e;
+            }
         }
 
         await writeBucketDeletedEvent({
